requirements.txt
apache-libcloud==2.1.0
asn1crypto==0.22.0
bcrypt==3.1.3
boto3==1.4.7
botocore==1.7.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
docutils==0.14
enum34==1.1.6
Fabric==1.14.0
future==0.16.0
futures==3.1.1
idna==2.6
ipaddress==1.0.18
jmespath==0.9.3
linode-api==4.1.2b0
paramiko==2.2.1
pyaml==17.8.0
pyasn1==0.3.3
pycparser==2.18
pycrypto==2.6.1
PyNaCl==1.1.2
python-dateutil==2.6.1
python-digitalocean==1.12
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.10
six==1.10.0
urllib3==1.22
